Big update
Moved stuff into a home directory to seperate it out.

Changed the names of the dotfiles to re-include the dot at the start of their
name.

Moved the scripts into an appropriate folder and updated the bootstrap script
to be more agnostic about where thigs are located.

Finally updated the submodule locations